junction_api
Module backend

Backends are the logical target of network traffic. They have an identity and a load-balancing policy. See Backend to get started.

Structs

Backend

A Backend is a logical target for network traffic.

BackendId

A Backend is uniquely identifiable by a combination of Service and port.

RequestHashPolicy

RingHashParams

Policy for configuring a ketama-style consistent hashing algorithm.

Enums

LbPolicy

A policy describing how traffic to this target should be load balanced.

RequestHasher