Module junction_api::http

HTTP routing configuration.

A Route specifies the all of the top-level configuration for all HTTP traffic that matches a particular URL.

Structs

HeaderFilter

Defines configuration for the RequestHeaderModifier filter.

HeaderValue

RequestMirrorFilter

Defines configuration for the RequestMirror filter.

RequestRedirectFilter

Defines a filter that redirects a request. This filter MUST not be used on the same Route rule as a URL Rewrite filter.

Route

High level policy that describes how a request to a specific hostname should be routed.

RouteMatch

Defines the predicate used to match requests to a given action. Multiple match types are ANDed together, i.e. the match will evaluate to true only if all conditions are satisfied.

RouteRetry

Specifies a way of configuring client retry policy.

RouteRule

Defines semantics for matching an HTTP request based on conditions (matches), processing it (filters), and forwarding the request to an API object (backendRefs).

RouteTimeouts

Defines timeouts that can be configured for a HTTP Route.

UrlRewriteFilter

Defines a filter that modifies a request during forwarding. At most one of these filters may be used on a Route rule. This may not be used on the same Route rule as a RequestRedirect filter.

WeightedTarget

Enums

HeaderMatch

Describes how to select a HTTP route by matching HTTP request headers.

PathMatch

Describes how to select a HTTP route by matching the HTTP request path.

PathModifier

Defines configuration for path modifiers.

QueryParamMatch

Describes how to select a HTTP route by matching HTTP query parameters.

RouteFilter

Defines processing steps that must be completed during the request or response lifecycle.

Type Aliases

HeaderName

The name of an HTTP header.

Method

Describes how to select a HTTP route by matching the HTTP method as defined by RFC 7231 and RFC 5789. The value is expected in upper case.