Module junction_api::shared

Shared configuration.

Structs

DNSTarget

Duration

A duration type where parsing and formatting obey the k8s Gateway API GEP-2257.

Fraction

A fraction, expressed as a numerator and a denominator.

Regex

A regular expression.

ServiceTarget

SessionAffinity

SessionAffinityHashParam

WeightedTarget

Enums

SessionAffinityHashParamType

Target

Type Aliases

PortNumber

Defines a network port.

PreciseHostname

The fully qualified domain name of a network host. This matches the RFC 1123 definition of a hostname with 1 notable exception that numeric IP addresses are not allowed.