junction_api

Module http

HTTP Route configuration. Routes dynamically congfigure things you might put directly in client code like timeouts and retries, failure detection, or picking a different backend based on request data.

Structs

BackendRef

HeaderFilter

Defines configuration for the RequestHeaderModifier filter.

HeaderValue

RequestMirrorFilter

Defines configuration for the RequestMirror filter.

RequestRedirectFilter

Defines a filter that redirects a request. This filter MUST not be used on the same Route rule as a URL Rewrite filter.

Route

A Route is a policy that describes how a request to a specific virtual host should be routed.

RouteMatch

Defines the predicate used to match requests to a given action. Multiple match types are ANDed together; the match will evaluate to true only if all conditions are satisfied. For example, if a match specifies a path match and two query_params matches, it will match only if the request’s path matches and both of the query_params are matches.

RouteRetry

Configure client retry policy.

RouteRule

A RouteRule contains a set of matches that define which requests it applies to, processing rules, and the final destination(s) for matching traffic.

RouteTimeouts

Defines timeouts that can be configured for a HTTP Route.

UrlRewriteFilter

Defines a filter that modifies a request during forwarding. At most one of these filters may be used on a Route rule. This may not be used on the same Route rule as a RequestRedirect filter.

Enums

HeaderMatch

Describes how to select a HTTP route by matching HTTP request headers.

HostnameMatch

A matcher for URL hostnames.

PathMatch

Describes how to select a HTTP route by matching the HTTP request path. The type of a match specifies how HTTP paths should be compared.

PathModifier

Defines configuration for path modifiers.

QueryParamMatch

Describes how to select a HTTP route by matching HTTP query parameters.

RouteFilter

Defines processing steps that must be completed during the request or response lifecycle.

Type Aliases

HeaderName

The name of an HTTP header.

Method

Describes how to select a HTTP route by matching the HTTP method as defined by RFC 7231 and RFC 5789. The value is expected in upper case.